member join service updated 10/19

+++ hotel_gabriella_project/hotel_gabriella_back/src/main/java/com/trip/hotel_gabriella/client/repository/JpaMemberRepository.java
@@ -18,10 +18,10 @@ public class JpaMemberRepository implements MemberRepository{
 
 
     @Override
-    public int findCountByAccount(String account) {
+    public Long findCountByAccount(String account) {
         String query = "SELECT COUNT(m.account) FROM Member m WHERE m.account = :account";
 
-        return em.createQuery(query, Integer.class)
+        return em.createQuery(query, Long.class)
                 .setParameter("account", account)
                 .getSingleResult();
     }
