+++ hotel_gabriella_project/hotel_gabriella_back/src/main/java/com/trip/hotel_gabriella/client/repository/MemberRepository.java
@@ -4,7 +4,7 @@ import com.trip.hotel_gabriella.common.domain.Member;
 
 public interface MemberRepository {
 
-    public int findCountByAccount(String account);
+    public Long findCountByAccount(String account);
 
     public void save(Member member);
